Move get_path_from_info() to strawberry.utils.info
